Move Notifications to shared folder
