+++ app/assets/styles/app.less
@@ -32,7 +32,6 @@
 
 // Components
 @import './loader';
-@import './notifications';
 @import './panel';
 @import './react-select';
 @import './reservation-info-modal';
