+++ app/containers/App.js
@@ -7,11 +7,11 @@ import { updatePath } from 'redux-simple-router';
 
 import { clearSearchResults } from 'actions/searchActions';
 import { fetchUser } from 'actions/userActions';
-import Notifications from 'containers/Notifications';
 import appSelector from 'selectors/containers/appSelector';
 import Favicon from 'shared/favicon';
 import Footer from 'shared/footer';
 import Navbar from 'shared/navbar';
+import Notifications from 'shared/notifications';
 import { getCustomizationClassName } from 'utils/customizationUtils';
 
 export class UnconnectedApp extends Component {
