+++ app/shared/notifications/index.js
@@ -1,0 +1,3 @@
+import NotificationsContainer from './NotificationsContainer';
+
+export default NotificationsContainer;
