+++ app/shared/shared.less
@@ -6,6 +6,7 @@
 @import './form-fields/form-fields';
 @import './modals/modals';
 @import './navbar/navbar';
+@import './notifications/notifications';
 @import './resource-icons/resource-icons';
 @import './resource-list/resource-list';
 @import './wrapped-text/wrapped-text';
